requirements.txt
pytorch
torchvision
PIL
pandas
prefetch_generator
tensorboardX
scipy